src/lambdas/periodic-screening-update-view/requirements.txt
boto3>=1.28.0
psycopg2-binary>=2.9.9